refactor: Position 객체 캐싱(Positions)
