+++ src/main/java/domain/position/Position.java
@@ -1,0 +1,35 @@
+package domain.position;
+
+import java.util.Objects;
+
+public final class Position {
+
+    private final File file;
+    private final Rank rank;
+
+    Position(final File file, final Rank rank) {
+        this.file = file;
+        this.rank = rank;
+    }
+
+    public String getName() {
+        return file.getName() + rank.getName();
+    }
+
+    @Override
+    public boolean equals(final Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (!(o instanceof Position)) {
+            return false;
+        }
+        final Position position = (Position) o;
+        return file == position.file && rank == position.rank;
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(file, rank);
+    }
+}
