+++ src/test/java/domain/board/BoardTest.java
@@ -9,6 +9,8 @@ import domain.piece.Pawn;
 import domain.piece.Piece;
 import domain.piece.Queen;
 import domain.piece.Rook;
+import domain.position.Position;
+import domain.position.Positions;
 import java.util.List;
 import java.util.Map;
 import org.junit.jupiter.api.DisplayName;
@@ -20,13 +22,13 @@ class BoardTest {
     void createTest() {
         //given
         Board board = Board.create();
-        final List<Position> rooksPosition = Position.of("A1", "A8", "H1", "H8");
-        final List<Position> knightsPosition = Position.of("B1", "B8", "G1", "G8");
-        final List<Position> bishopsPosition = Position.of("C1", "C8", "F1", "F8");
-        final List<Position> pawnsPosition = Position.of("A2", "B2", "C2", "D2", "E2", "F2", "G2", "H2",
+        final List<Position> rooksPosition = Positions.of("A1", "A8", "H1", "H8");
+        final List<Position> knightsPosition = Positions.of("B1", "B8", "G1", "G8");
+        final List<Position> bishopsPosition = Positions.of("C1", "C8", "F1", "F8");
+        final List<Position> pawnsPosition = Positions.of("A2", "B2", "C2", "D2", "E2", "F2", "G2", "H2",
                 "A7", "B7", "C7", "D7", "E7", "F7", "G7", "H7");
-        final List<Position> kingsPosition = Position.of("E1", "E8");
-        final List<Position> queensPosition = Position.of("D1", "D8");
+        final List<Position> kingsPosition = Positions.of("E1", "E8");
+        final List<Position> queensPosition = Positions.of("D1", "D8");
 
         //when
         final Map<Position, Piece> pieces = board.getPieces();
