+++ src/test/java/domain/board/PositionTest.java
@@ -4,10 +4,14 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertThrowsExactly;
 
+import domain.position.Position;
+import domain.position.Positions;
 import java.util.List;
 import java.util.stream.Collectors;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ValueSource;
 
 class PositionTest {
     @DisplayName("문자열을 받아 Position을 생성할 수 있다.")
@@ -19,20 +23,20 @@ class PositionTest {
         //when
 
         //then
-        assertDoesNotThrow(() -> Position.from(position));
+        assertDoesNotThrow(() -> Positions.from(position));
     }
 
     @DisplayName("체스판 범위를 벗어난 문자열을 받으면 예외가 발생한다.")
-    @Test
-    void createFail() {
+    @ParameterizedTest
+    @ValueSource(strings = {"Z1", "AB5", ""})
+    void createFail(String position) {
         //given
-        final String position = "Z1";
 
         //when
 
         //then
         assertThrowsExactly(IllegalArgumentException.class,
-                () -> Position.from(position));
+                () -> Positions.from(position));
     }
 
     @DisplayName("위치를 여러 개를 받아서 만들 수 있다.")
@@ -41,13 +45,13 @@ class PositionTest {
         //given
         final List<String> positions = List.of("A3", "H2", "G5");
         final List<Position> expected = positions.stream()
-                .map(Position::from)
+                .map(Positions::from)
                 .collect(Collectors.toList());
 
         //when
 
         //then
-        assertThat(Position.of(positions.toArray(String[]::new))).isEqualTo(expected);
+        assertThat(Positions.of(positions.toArray(String[]::new))).isEqualTo(expected);
     }
 }
